+++ common/src/main/java/com/nb6868/onex/common/filter/JwtTokenFilter.java
@@ -26,6 +26,7 @@ import java.io.IOException;
  *
  * @author Charles zhangchaoxu@gmail.com
  */
+@Deprecated
 public class JwtTokenFilter implements Filter {
 
     /**